pyproject.toml
[project]
name = "llm-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crewai>=0.95.0",
    "google-adk>=1.15.0",
    "langchain>=0.3.27",
    "langchain-openai>=0.3.33",
    "langgraph>=0.6.7",
    "litellm>=1.77.4",
    "python-dotenv>=1.1.1",
]
